Added Score handling script
Updated the obstacle movement script, when an obstacle moves past the player it counts as a single point, this is determined using the x coordinate and checking if it is less than 0 which is the players location.

Bug introduced - craps and starfish contribute to the score which i do not want.

Added code to ScoreKeeper script and removed the update function which i was using for debugging purposes